Visualize Proportions in Bar Plot

This function creates a bar plot to visualize the proportions of a specified group within the data. The bars are colored according to a specified metadata column and can be grouped by another metadata column.

Usage

plotBar(data, color_by = NULL, split_by = NULL)

Arguments

data

A Cygnus object containing EV metadata and matrices.

color_by

A column name from the EV metadata to use for coloring the proportions in the plot. This defines the different groups/colors in the bar plot.

split_by

A column name from the EV metadata that defines how to group the data on the x-axis.

Value

A ggplot2 object representing the bar plot of proportions.