Plot Distributions of Selected Markers

This function plots the distribution of expression levels for selected markers in an expression matrix. The user can specify a subset of markers to plot, or plot all markers by default.

Usage

plotDistribution(obj, plot_markers = "ALL", matrix = "Raw_Score")

Arguments

obj

An object of class CygnusObject containing the expression matrix.

plot_markers

A vector of marker names to plot. If set to "ALL", distributions for all markers will be plotted. Default is "ALL".

matrix

Character string specifying the name of the matrix to use. Default is "Raw_Score".

Value

A series of histogram plots, each representing the distribution of expression levels for the specified markers.