Generate an UpSet Plot from Colocalization Statistics

Generate an UpSet Plot from Colocalization Statistics

Usage

plotUpset(
  data,
  colocal_data,
  matrix_name = "binary_exp_mat",
  threshold_count = 1,
  min_degree_setting = 1
)

Arguments

data

An object of class CygnusObject containing the binary expression matrix.

colocal_data

A data frame of deviations and statistics from running getColocalizedMarkers

matrix_name

The matrix within the data object that contains the binary expression matrix. The default is set to binary_exp_mat

threshold_count

An integer indicating the minimum number of elements in an intersection

min_degree_setting

An integer indicating the minimum degree of intersection for plotting