Perform UMAP on a Specified Matrix and Store in Dimensionality Reduction Slot

Perform UMAP on a Specified Matrix and Store in Dimensionality Reduction Slot

Usage

runUMAP(
  data,
  matrix_name = "Raw_Score",
  use_relevant = TRUE,
  dims = 3,
  use_pcs = FALSE,
  n_pcs = 5,
  n_neighbors = 30,
  min_dist = 0.3,
  verbose = TRUE,
  ...
)

Arguments

data

An object of class CygnusObject containing an expression matrix.

matrix_name

Character string specifying the name of the matrix to use. Default is "Raw_Score".

use_relevant

Logical indicating whether to use only relevant markers. Default is TRUE.

dims

Numeric value specifying the number of dimensions for UMAP. Default is 2.

use_pcs

Logical indicating whether to use PCA results as input. Default is FALSE.

n_pcs

Number of principal components. Used only when use_pcs = TRUE. Default is 5.

n_neighbors

Numeric value specifying the number of neighbors for UMAP. Default is 30.

min_dist

Numeric value specifying the minimum distance for UMAP. Default is 0.3.

verbose

Logical value indicating whether to display a progress bar. Default is TRUE.

...

Additional arguments to pass to the uwot::umap function.

Value

The updated CygnusObject with UMAP coordinates stored in the 'dim_red' slot.